cli cleanup
